Result added to PrivateLessonService

+++ Find-A-Tutor.Core/Services/PrivateLessonService.cs
@@ -1,8 +1,6 @@
 ﻿using AutoMapper;
 using Find_A_Tutor.Core.Domain;
 using Find_A_Tutor.Core.DTO;
-using Find_A_Tutor.Core.Exceptions;
-using Find_A_Tutor.Core.Extensions;
 using Find_A_Tutor.Core.Repositories;
 using NLog;
 using System;
@@ -37,51 +35,69 @@ namespace Find_A_Tutor.Core.Services
                                     Result<PrivateLessonDTO>.Error($"Private lesson with id: {id}, does not exists.");
         }
 
-        public async Task<PrivateLessonDTO> GetAsyncBySubject(string name)
+        public async Task<Result<PrivateLessonDTO>> GetAsyncBySubject(string subject)
         {
-            var privateLesson = await _privateLessonRepository.GetAsyncBySubject(name);
-            return _mapper.Map<PrivateLessonDTO>(privateLesson);
+            var privateLesson = await _privateLessonRepository.GetAsyncBySubject(subject);
+            return privateLesson != null ?
+                                    Result<PrivateLessonDTO>.Ok(_mapper.Map<PrivateLessonDTO>(privateLesson)) :
+                                    Result<PrivateLessonDTO>.Error($"There are no private lessons that subject contains \"{subject}\".");
         }
 
-        public async Task<IEnumerable<PrivateLessonDTO>> GetForUserAsync(Guid userId)
+        public async Task<Result<IEnumerable<PrivateLessonDTO>>> GetForUserAsync(Guid userId)
         {
-            var user = await _userRepository.GetOrFailAsync(userId);
+            var user = await _userRepository.GetAsync(userId);
+            if (user == null)
+            {
+                return Result<IEnumerable<PrivateLessonDTO>>.Error($"User with id: '{userId}' does not exist.");
+            }
             var allLessons = await _privateLessonRepository.BrowseAsync();
 
-            var allLessonsForUser = new List<PrivateLessonDTO>();
+            var allLessonsForUser = _mapper.Map<IEnumerable<PrivateLessonDTO>>(allLessons.Where(x => x.StudentId == userId || x.TutorId == userId));
 
-            allLessonsForUser.AddRange(_mapper.Map<IEnumerable<PrivateLessonDTO>>(allLessons.Where(x => x.StudentId == userId || x.TutorId == userId)));
-
-            return allLessonsForUser;
+            return Result<IEnumerable<PrivateLessonDTO>>.Ok(allLessonsForUser);
         }
 
-        public async Task<IEnumerable<PrivateLessonDTO>> BrowseAsync(string description = "")
+        public async Task<Result<IEnumerable<PrivateLessonDTO>>> BrowseAsync(string description = "")
         {
             logger.Info("Fetching events");
             var privateLesson = await _privateLessonRepository.BrowseAsync(description);
-            return _mapper.Map<IEnumerable<PrivateLessonDTO>>(privateLesson);
+            return Result<IEnumerable<PrivateLessonDTO>>.Ok(_mapper.Map<IEnumerable<PrivateLessonDTO>>(privateLesson));
         }
 
-        public async Task CreateAsync(Guid id, Guid studnetId, DateTime relevantTo, string description, string subject)
+        public async Task<Result> CreateAsync(Guid id, Guid studnetId, DateTime relevantTo, string description, string subject)
         {
             var privateLesson = await _privateLessonRepository.GetAsync(id);
             if (privateLesson != null)
             {
-                throw new RepositoryException($"Private lesson already exists.");
+                return Result.Error("Private lesson already exists.");
             }
 
-            var schoolSubject = await _schoolSubjectRepository.GetOrFailAsync(subject);
+            var schoolSubject = await _schoolSubjectRepository.GetAsync(subject);
+            if (schoolSubject == null)
+            {
+                return Result.Error($"School subject with name: '{subject}' does not exist.");
+            }
 
             privateLesson = new PrivateLesson(id, studnetId, relevantTo, description, schoolSubject);
             await _privateLessonRepository.AddAsync(privateLesson);
 
-            logger.Info($"Private lesson with id: {0}, was successfully created", id);
+            logger.Info($"Private lesson with id: {id}, was successfully created");
+            return Result.Ok();
         }
 
-        public async Task UpdateAsync(Guid privateLessonId, DateTime relevantTo, string description, string subject)
+        public async Task<Result> UpdateAsync(Guid privateLessonId, DateTime relevantTo, string description, string subject)
         {
-            var privateLesson = await _privateLessonRepository.GetOrFailAsync(privateLessonId);
-            var schoolSubject = await _schoolSubjectRepository.GetOrFailAsync(subject);
+            var privateLesson = await _privateLessonRepository.GetAsync(privateLessonId);
+            if (privateLesson == null)
+            {
+                return Result.Error($"Private lesson with id: '{privateLessonId}' does not exist.");
+            }
+
+            var schoolSubject = await _schoolSubjectRepository.GetAsync(subject);
+            if (schoolSubject == null)
+            {
+                return Result.Error($"Private lesson with name: '{subject}' does not exist.");
+            }
 
             privateLesson.SetDesctiption(description);
             privateLesson.SetRelevantToDate(relevantTo);
@@ -89,40 +105,63 @@ namespace Find_A_Tutor.Core.Services
 
             await _privateLessonRepository.UpdateAsync(privateLesson);
 
-            logger.Info($"Private lesson with id: {0}, was successfully updated", privateLessonId);
+            logger.Info($"Private lesson with id: {privateLessonId}, was successfully updated");
+            return Result.Ok();
         }
 
-        public async Task DeleteAsync(Guid privateLessonId)
+        public async Task<Result> DeleteAsync(Guid privateLessonId)
         {
-            var privateLesson = await _privateLessonRepository.GetOrFailAsync(privateLessonId);
+            var privateLesson = await _privateLessonRepository.GetAsync(privateLessonId);
+            if (privateLesson == null)
+            {
+                return Result.Error($"Private lesson with id: '{privateLessonId}' does not exist.");
+            }
             await _privateLessonRepository.DeleteAsync(privateLesson);
 
-            logger.Info($"Private lesson with id: {0}, was successfully deleted", privateLessonId);
+            logger.Info($"Private lesson with id: {privateLessonId}, was successfully deleted");
+            return Result.Ok();
         }
 
-        public async Task AssignTutor(Guid privateLessonId, Guid tutorId)
+        public async Task<Result> AssignTutor(Guid privateLessonId, Guid tutorId)
         {
-            var privateLesson = await _privateLessonRepository.GetOrFailAsync(privateLessonId);
+            var privateLesson = await _privateLessonRepository.GetAsync(privateLessonId);
+            if (privateLesson == null)
+            {
+                return Result.Error($"Private lesson with id: '{privateLessonId}' does not exist.");
+            }
+
             if (privateLesson.TutorId != null)
             {
-                throw new RepositoryException($"Private lesson is already assigned.");
+                return Result.Error($"Private lesson is already assigned.");
             }
-            var tutor = await _userRepository.GetOrFailAsync(tutorId);
+            var tutor = await _userRepository.GetAsync(tutorId);
+            if (tutor == null)
+            {
+                return Result.Error($"User with id: '{tutorId}' does not exist.");
+            }
+
             privateLesson.AssignTutor(tutor);
 
-            logger.Info($"Tutor with id {0} was assigned to lesson with id {1}", tutorId, privateLessonId);
+            logger.Info($"Tutor with id {tutorId} was assigned to lesson with id {privateLessonId}");
+            return Result.Ok();
         }
 
-        public async Task RemoveAssignedTutor(Guid privateLessonId, Guid userId)
+        public async Task<Result> RemoveAssignedTutor(Guid privateLessonId, Guid userId)
         {
-            var privateLesson = await _privateLessonRepository.GetOrFailAsync(privateLessonId);
+            var privateLesson = await _privateLessonRepository.GetAsync(privateLessonId);
+            if (privateLesson == null)
+            {
+                return Result.Error($"Private lesson with id: '{privateLessonId}' does not exist.");
+            }
+
             if (privateLesson.TutorId != userId && privateLesson.StudentId != userId)
             {
-                throw new RepositoryException("Tutor can be unassign only by orignal student or tutor.");
+                return Result.Error("Tutor can be unassign only by orignal student or tutor.");
             }
             privateLesson.RemoveAssignedTutor();
 
-            logger.Info($"Assigned tutor was removed from lesson with id {1}", privateLessonId);
+            logger.Info($"Assigned tutor was removed from lesson with id {privateLessonId}");
+            return Result.Ok();
         }
     }
 }
